Fix: restructure the resolvers structure for a better implementation using business segment to separate each group of resolvers
